WIP migrate search query composition from utilize_indices to a utility file
