+++ src/execution_plan/ops/op_index_scan.h
@@ -17,14 +17,15 @@ typedef struct {
 	OpBase op;
 	Graph *g;
 	RSIndex *idx;
-	NodeScanCtx n;           /* Label data of node being scanned. */
-	uint nodeRecIdx;            /* Index of the node being scanned in the Record. */
-	RSQNode *rs_query_node;     /* RediSearch query node used to construct iterator. */
-	RSResultsIterator *iter;    /* RediSearch iterator over an index with the appropriate filters. */
-	Record child_record;        /* The Record this op acts on if it is not a tap. */
+	const FT_FilterNode *filter;
+	NodeScanCtx n;             // label data of node being scanned
+	uint nodeRecIdx;           // index of the node being scanned in the Record
+	RSQNode *rs_query_node;    // rediSearch query node used to construct iterator
+	RSResultsIterator *iter;   // rediSearch iterator over an index with the appropriate filters
+	Record child_record;       // the Record this op acts on if it is not a tap
 } IndexScan;
 
-/* Creates a new IndexScan operation */
+// creates a new IndexScan operation
 OpBase *NewIndexScanOp(const ExecutionPlan *plan, Graph *g, NodeScanCtx n, RSIndex *idx,
-					   RSQNode *rs_query_node);
+					   const FT_FilterNode *filter);
